core: speed up QCDM probing a bit
The point of sending two "version info" commands was to ensure that
the terminating 0x7E of the first one was processed as a QCDM frame
boundary and that any random data in the buffer (like AT commands
from probing) got cleared out.  The second command would always
get processed as a valid QCDM command if the device supported QCDM,
since there was no garbage before it.

Instead of that dance, just prepend the version info message with
an extra 0x7E to ensure a clean QCDM frame which the device hopefully
responds to immediately.  Second, actually process that response
instead of throwing it away.  Should save about 3 seconds when
probing QCDM ports.