sierra: fix CSQ handling on APP1 port
The APP1 port doesn't always prefix its replies with <CR><LF> which
runs afoul of the built-in echo removal.  Since Sierra modems are on
the whole well-behaved WRT echo removal, just disable it on the
secondary ports.  Only changes behavior for PPP-based devices since
they are the only ones that use the APP1 ports.