broadband-modem-qmi: don't always trust the current capabilities reported by SSP
Sometimes the 'Mode Preference' TLV in the SSP response will have incorrect
values. The following log is for a Gobi3k modem which is *only* GMS/UMTS.

ModemManager[13415]: [/dev/cdc-wdm0] Received message...
>>>>>> QMUX:
>>>>>>   length  = 44
>>>>>>   flags   = 0x80
>>>>>>   service = "nas"
>>>>>>   client  = 2
>>>>>> QMI:
>>>>>>   flags       = "response"
>>>>>>   transaction = 1
>>>>>>   tlv_length  = 32
>>>>>>   message     = "Get System Selection Preference" (0x0034)
>>>>>> TLV:
>>>>>>   type       = "Result" (0x02)
>>>>>>   length     = 4
>>>>>>   value      = 00:00:00:00
>>>>>>   translated = SUCCESS
>>>>>> TLV:
>>>>>>   type       = "Emergency mode" (0x10)
>>>>>>   length     = 1
>>>>>>   value      = 00
>>>>>>   translated = 0
>>>>>> TLV:
>>>>>>   type       = "Mode Preference" (0x11)
>>>>>>   length     = 2
>>>>>>   value      = 0F:00
>>>>>>   translated = 15
>>>>>> TLV:
>>>>>>   type       = "Band Preference" (0x12)
>>>>>>   length     = 8
>>>>>>   value      = FF:FF:FF:3F:FF:FF:FF:FF
>>>>>>   translated = 18446744070488326143
>>>>>> TLV:
>>>>>>   type       = "Roaming Preference" (0x14)
>>>>>>   length     = 2
>>>>>>   value      = FF:00
>>>>>>   translated = 255

Instead of setting to '1' only the bits related to GSM/UMTS in the "Mode
Preference" TLV, the modem sets all the bits it knows of to '1', including the
CDMA/EVDO ones. It is not easy to detect this properly, because the modem didn't
set to '1' the bits it didn't know of (e.g. LTE or TD-SCDMA), so we *cannot*
assume that 0x000F actually means 'unknown'.

So, we now will pass these returned values through the "DMS Get Capabilities"
filter, as we did for the Technology Preference result. The final reported
current capabilities will be either:
 * The intersection between the capabilities loaded by "NAS Get System Selection
   Preference" and "DMS Get Capabilities" or,
 * The intersection between the capabilities loaded by "NAS Get Technology
   Preference" and "DMS Get Capabilities".

For 'modem' (not 'current') capabilities, we still rely only on "DMS Get
Capabilities".