sierra: fix CFUN power up delay handling
1) all Sierra devices appear to require short delay after powering up,
otherwise subsequent commands may return errors.  Older devices need
longer so ensure new devices are penalized just for being new.

2) When the modem is already in full functionality status and no power
up command was sent, there's no need to delay, which was happening
regardless of what state the modem was already in.  Detect whether
the power up was actually executed (response and error will be NULL)
and only delay if it was executed.