libmm-glib,manager: register DBus error associations in advance
In order to have a proper conversion between DBus error names and GErrors for
our known domains, the associations need to be registered before any DBus
call attempt.

Given that the high-level API of libmm-glib has its entry point always in the
MMManager, just register them as soon as the first such object is created.