gsm: always try +CSQ if +CIND fails; fix bugs
Two issues:

1) if there was a generic error with the +CIND response, the
code set the callback error, scheduled the callback, but then
continued with +CIND and possibly +CSQ.  That could overwrite
the original error (a memory leak) and trigger warning about an
already-scheduled callback.  Instead, just ignore the error and
request +CSQ as a fallback.

2) If +CIND parsing failed, the parse error was set on the
callback, but +CSQ was also requested.  Thus if +CSQ was
successful, the +CIND error would still be returned.  Instead,
just ignore the +CIND parsing error (but log it) and request
+CSQ as a fallback.