gsm,cdma: don't reset connected state if disconnecting fails
If disconnecting fails, we can't be sure we're not still connected,
which is why previously MM reset the CONNECTED state.  However, this
 behavior in conjunction with 27408b4c1e13e3016a01e5b273e66d99f487d3ad
caused ModemManager not to recognize that Bluetooth rfcomm ports have
disappeared.  Oddly udev doesn't tell us that they have gone away,
probably because ModemManager still has the port's file descriptor
open and virtual devices like rfcomm ports sometimes stick around
until all their users have disappeared.  Because of 27408b4c MM
wasn't listening to the serial port and the port hangup that would
normally cause MM to release the port was ignored.

But in the end, resetting CONNECTED state when disconnecting doesn't
do anything useful because it doesn't allow commands to be sent to
the modem which might (helpfully) interrupt the command stream and
get the device responsive.  Also, whatever terminated the data
connection and called Disconnect() should already have terminated
anything that would allow data to be transferred, like the PPP
session or cleared the IP addresses from the net interface, or
terminated the Bluetooth link.