icera: fix issues checking supported bands
Icera devices include bands that the modem doesn't support in
the %IPBM=? list, so the plugin sets the band to its current
enabled/disabled value to test whether that band is supported.

Apparently enabling a band that's already enabled isn't a NOP,
and we don't want to make any changes since we're just checking
whether the band is supported.  So instead of enabling already-
enabled bands, just assume those bands are supported and only
check the disabled bands.

This is a backport of 23204111130fca0a93a22508431ad8bb17210bdd